requirements.txt
python-telegram-bot==21.6
anthropic>=0.40.0
python-dotenv==1.0.1
supabase==2.3.0
httpx==0.27.2
